ocr_service/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
pillow-heif==0.18.0
Pillow==10.4.0
